# docs/openapi.yaml
openapi: 3.1.0
info:
    title: RSTC API
    version: 1.0.0
    description: REST API for RSTC application.
servers:
    - url: /api
paths:
...
    /missions:
        post:
            tags:
                - Missions
            summary: Create mission
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/MissionCreate'
                        examples:
                            create:
                                value:
                                    name: ماموریت استان تهران
                                    lname: رضایی
                                    start_date: '1403-07-01'
                                    end_date: '1403-07-05'
                                    issue_date: '1403-06-28'
                                    mission_type: Repair
                                    region: تهران
                                    location: مرکزی
                                    subject: تعمیر تجهیزات
                                    device_type: Laptop
                                    device_serial: ABC123
                                    duration: 5
                                    overtime_hours: 0
                                    is_single: true
                                    is_group: false
                                    is_supplied: true
                                    is_unsupplied: false
                                    is_issued: false
                                    is_extended: false
                                    is_gov: false
                                    is_plane: false
                                    is_train: false
                                    is_agency: false
                                    is_bus: false
                                    is_personal: false
            responses:
                '200':
                    description: Mission created
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                    decree_num:
                                        type: string
                                        example: RSTC-2026-07-27-0001
                        examples:
                            success:
                                value:
                                    success: true
                                    decree_num: RSTC-2026-07-27-0001
                '400':
                    $ref: '#/components/responses/Error'
                '500':
                    $ref: '#/components/responses/Error'
        get:
            tags:
                - Missions
            summary: List missions
            security:
                - bearerAuth: []
            responses:
                '200':
                    description: Missions list
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Mission'
                '500':
                    $ref: '#/components/responses/Error'
    /missions/{id}:
        put:
            tags:
                - Missions
            summary: Update mission
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/MissionUpdate'
                        examples:
                            update:
                                value:
                                    name: ماموریت به روز شده
                                    start_date: '1403-07-02'
                                    end_date: '1403-07-06'
                                    issue_date: '1403-06-29'
            responses:
                '200':
                    description: Mission updated
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessResponse'
                '400':
                    $ref: '#/components/responses/Error'
                '500':
                    $ref: '#/components/responses/Error'
        delete:
            tags:
                - Missions
            summary: Delete mission
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
            responses:
                '200':
                    description: Mission deleted
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessResponse'
                '500':
                    $ref: '#/components/responses/Error'
        get:
            tags:
                - Missions
            summary: PDF export (client-side only)
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
            responses:
                '404':
                    description: PDF export is client-side only
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                            examples:
                                notFound:
                                    value:
                                        error: 'PDF export is now client-side only. Use jsPDF for PDF generation.'
    /reports/missions:
        post:
            tags:
                - Reports
            summary: Search missions
            security:
                - bearerAuth: []
            requestBody:
                required: false
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                lname:
                                    type: string
                                emp_num:
                                    type: string
                                decree_num:
                                    type: string
                                region:
                                    type: string
                                mission_type:
                                    type: string
                                location:
                                    type: string
                                start_date_from:
                                    type: string
                                start_date_to:
                                    type: string
                                end_date_from:
                                    type: string
                                end_date_to:
                                    type: string
                                issue_date_from:
                                    type: string
                                issue_date_to:
                                    type: string
                        examples:
                            filter:
                                value:
                                    region: تهران
                                    mission_type: Repair
            responses:
                '200':
                    description: Search results
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    results:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/Mission'
                                    total:
                                        type: integer
                            examples:
                                result:
                                    value:
                                        results: []
                                        total: 0
                '500':
                    $ref: '#/components/responses/Error'
...
    /backup:
        get:
            tags:
                - Backup
            summary: Download database backup
            security:
                - bearerAuth: []
            responses:
                '200':
                    description: Database backup file
                    content:
                        application/octet-stream:
                            schema:
                                type: string
                                format: binary
                            examples:
                                backup:
                                    value: 'binary data'
                '404':
                    $ref: '#/components/responses/Error'
                '500':
                    $ref: '#/components/responses/Error'
    /backups:
        get:
            tags:
                - Backup
            summary: List backup files
            security:
                - bearerAuth: []
            responses:
                '200':
                    description: Backup files list
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    backups:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/Backup'
                                    settings:
                                        type: object
                                        properties:
                                            maxBackups:
                                                type: integer
                                            scheduleHour:
                                                type: integer
                            examples:
                                list:
                                    value:
                                        backups: []
                                        settings:
                                            maxBackups: 30
                                            scheduleHour: 2
                '500':
                    $ref: '#/components/responses/Error'
    /backups/{name}:
        get:
            tags:
                - Backup
            summary: Download specific backup file
            security:
                - bearerAuth: []
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                '200':
                    description: Backup file
                    content:
                        application/octet-stream:
                            schema:
                                type: string
                                format: binary
                '404':
                    $ref: '#/components/responses/Error'
                '500':
                    $ref: '#/components/responses/Error'
        delete:
            tags:
                - Backup
            summary: Delete backup file
            security:
                - bearerAuth: []
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                '200':
                    description: Backup deleted
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                    message:
                                        type: string
                            examples:
                                deleted:
                                    value:
                                        success: true
                                        message: 'فایل پشتیبان حذف شد.'
                '404':
                    $ref: '#/components/responses/Error'
                '500':
                    $ref: '#/components/responses/Error'
    /backups/validate:
        post:
            tags:
                - Backup
            summary: Validate backup file
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/octet-stream:
                        schema:
                            type: string
                            format: binary
            responses:
                '200':
                    description: Validation result
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    valid:
                                        type: boolean
                                    sizeMB:
                                        type: string
                                    tables:
                                        type: array
                                        items:
                                            type: string
                                    counts:
                                        type: object
                                    integrity:
                                        type: string
                                    pageCount:
                                        type: integer
                                    pageSize:
                                        type: integer
                                    estimatedSizeMB:
                                        type: string
                            examples:
                                valid:
                                    value:
                                        valid: true
                                        sizeMB: '12.34'
                                        tables:
                                            - Personnel
                                            - Missions
                                        counts:
                                            Personnel: 100
                                            Missions: 20
                                        integrity: ok
                                        pageCount: 500
                                        pageSize: 4096
                                        estimatedSizeMB: '1.95'
                '400':
                    $ref: '#/components/responses/Error'
                '500':
                    $ref: '#/components/responses/Error'
    /restore:
        post:
            tags:
                - Backup
            summary: Restore database backup
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/octet-stream:
                        schema:
                            type: string
                            format: binary
            responses:
                '200':
                    description: Restore successful
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                    message:
                                        type: string
                            examples:
                                restored:
                                    value:
                                        success: true
                                        message: 'بازیابی با موفقیت انجام شد و اطلاعات جدید فعال شدند.'
                '500':
                    $ref: '#/components/responses/Error'
...
components:
    schemas:
...
        Mission:
            type: object
            properties:
                id:
                    type: integer
                decree_num:
                    type: string
                name:
                    type: string
                lname:
                    type: string
                emp_num:
                    type: string
                job_title:
                    type: string
                mission_type:
                    type: string
                device_type:
                    type: string
                repair_type:
                    type: string
                region:
                    type: string
                location:
                    type: string
                subject:
                    type: string
                device_serial:
                    type: string
                duration:
                    type: integer
                overtime_hours:
                    type: integer
                start_date:
                    type: string
                end_date:
                    type: string
                issue_date:
                    type: string
                is_single:
                    type: boolean
                is_group:
                    type: boolean
                is_supplied:
                    type: boolean
                is_unsupplied:
                    type: boolean
                is_issued:
                    type: boolean
                is_extended:
                    type: boolean
                is_gov:
                    type: boolean
                is_plane:
                    type: boolean
                is_train:
                    type: boolean
                is_agency:
                    type: boolean
                is_bus:
                    type: boolean
                is_personal:
                    type: boolean
        MissionCreate:
            type: object
            required:
                - name
                - start_date
                - end_date
                - issue_date
            properties:
                name:
                    type: string
                lname:
                    type: string
                emp_num:
                    type: string
                job_title:
                    type: string
                mission_type:
                    type: string
                device_type:
                    type: string
                repair_type:
                    type: string
                region:
                    type: string
                location:
                    type: string
                subject:
                    type: string
                device_serial:
                    type: string
                duration:
                    type: integer
                overtime_hours:
                    type: integer
                start_date:
                    type: string
                end_date:
                    type: string
                issue_date:
                    type: string
                is_single:
                    type: boolean
                is_group:
                    type: boolean
                is_supplied:
                    type: boolean
                is_unsupplied:
                    type: boolean
                is_issued:
                    type: boolean
                is_extended:
                    type: boolean
                is_gov:
                    type: boolean
                is_plane:
                    type: boolean
                is_train:
                    type: boolean
                is_agency:
                    type: boolean
                is_bus:
                    type: boolean
                is_personal:
                    type: boolean
        MissionUpdate:
            type: object
            required:
                - name
                - start_date
                - end_date
                - issue_date
            properties:
                name:
                    type: string
                lname:
                    type: string
                emp_num:
                    type: string
                job_title:
                    type: string
                mission_type:
                    type: string
                device_type:
                    type: string
                repair_type:
                    type: string
                region:
                    type: string
                location:
                    type: string
                subject:
                    type: string
                device_serial:
                    type: string
                duration:
                    type: integer
                overtime_hours:
                    type: integer
                start_date:
                    type: string
                end_date:
                    type: string
                issue_date:
                    type: string
                is_single:
                    type: boolean
                is_group:
                    type: boolean
                is_supplied:
                    type: boolean
                is_unsupplied:
                    type: boolean
                is_issued:
                    type: boolean
                is_extended:
                    type: boolean
                is_gov:
                    type: boolean
                is_plane:
                    type: boolean
                is_train:
                    type: boolean
                is_agency:
                    type: boolean
                is_bus:
                    type: boolean
                is_personal:
                    type: boolean
...
        Backup:
            type: object
            properties:
                name:
                    type: string
                size:
                    type: integer
                sizeMB:
                    type: string
                modified:
                    type: string
                    format: date-time
                modifiedJalali:
                    type: string
...
        SuccessResponse:
            type: object
            properties:
                success:
                    type: boolean
        ErrorResponse:
            type: object
            properties:
                error:
                    type: string
    responses:
        Error:
            description: Error response
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    examples:
                        badRequest:
                            value:
                                error: 'Invalid request'
                        unauthorized:
                            value:
                                error: 'دسترسی غیرمجاز! لطفاً ابتدا وارد شوید.'
                        forbidden:
                            value:
                                error: 'دسترسی غیرمجاز'
                        notFound:
                            value:
                                error: 'یافت نشد'
                        serverError:
                            value:
                                error: 'خطای سرور'